+++ src/main/scala/ai/starlake/extractor/ExtractScriptGenConfig.scala
@@ -28,7 +28,7 @@ object ExtractScriptGenConfig extends CliConfig[ExtractScriptGenConfig] {
       note(
         """
           |For domain extraction, the schemas should at least, specify :
-          |- a table name (schemas.renameSource else schemas.name)
+          |- a table name (schemas.name)
           |- a file pattern (schemas.pattern) which is used as the export file base name
           |- a write mode (schemas.metadata.write): APPEND or OVERWRITE
           |- a delta column (schemas.merge.timestamp) if in APPEND mode : the default column which is used to determine new rows for each exports