+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 
 # 0.3.13
 __New feature__:
-- Improve XLS support for schemas with a new field: rename_source
+- Improve XLS support for schemas with a long lame > 31 chars
 
 # 0.3.12
 __Bug Fix__: reorder transformation on attributes as follows: